Write class declaration Grub.js in ES6
Things to note:
the 'constructor' keyword. It seems that
within the constructor, the desired
properties are appended to the object
we are instantiating.

Also methods are declared within the the
scope of the class declaration.  It seems
to imply that as long as the the method
is not declared within the constructor,
the method is appended to the prototype.