Write class declaration subclass Bee.js
Things to remember:
(1) The super() method is the equivalent
to SuperClass.call(this).  And this must
be called before properties or methods
are declared on the subclass.

(2)extends key word is the equivalent to
Subclass.prototype = Object.create(Superclass.
prototype); Seems to be that extends will
point the __proto__ to an instance of the
another objects prototype that is specified.